This (hopefully) finishes this PR. Last commits have added the flushing of states to a binary file and flushing of routes to a json file.' '

+++ model/parts/workers/routing_worker.go
@@ -45,7 +45,7 @@ func RoutingWorker(requestChan chan types.Request, routeChan chan types.RouteDat
 			if curTimeStep%1000000 == 0 {
 				fmt.Println("routeChan: ", len(routeChan))
 			}
-			routeChan <- types.RouteData{TimeStep: curTimeStep, Route: route}
+			routeChan <- types.RouteData{TimeStep: int32(curTimeStep), Route: route}
 		}
 
 		if constants.Constants.IsWriteStatesToFile() {
